scripts/fetch_nq_data.py
#!/usr/bin/env python3
"""
NQ Futures Data Fetcher - Starter script for fetching NQ futures data.
Usage: python scripts/fetch_nq_data.py
"""
import os
import logging
from datetime import datetime
from pathlib import Path
import pandas as pd
import yaml

logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
logger = logging.getLogger("nq_data_fetcher")
ROOT_DIR = Path(__file__).resolve().parent.parent
CONFIG_PATH = ROOT_DIR / "config" / "settings.yaml"

def load_config():
    config_path = CONFIG_PATH if CONFIG_PATH.exists() else ROOT_DIR / "config" / "settings.example.yaml"
    with open(config_path, "r") as f:
        return yaml.safe_load(f)

def fetch_nq_daily_ohlcv(start_date, end_date, symbol="NQ"):
    logger.info(f"Fetching {symbol} data from {start_date} to {end_date}")
    # TODO: Replace with actual data source (yfinance, CME DataMine, etc.)
    df = pd.DataFrame(columns=["date", "open", "high", "low", "close", "volume"])
    logger.warning("Placeholder - configure a data source in this script.")
    return df

def main():
    config = load_config()
    start = config.get("date_range", {}).get("start_date", "2020-01-01")
    end = config.get("date_range", {}).get("end_date", "2026-12-31")
    symbol = config.get("contract", {}).get("symbol", "NQ")
    df = fetch_nq_daily_ohlcv(start, end, symbol)
    if not df.empty:
        raw_path = ROOT_DIR / config.get("storage", {}).get("raw_data_path", "data/raw")
        raw_path.mkdir(parents=True, exist_ok=True)
        filepath = raw_path / f"nq_daily_ohlcv_{datetime.now().strftime('%Y%m%d')}.csv"
        df.to_csv(filepath, index=False)
        logger.info(f"Saved {len(df)} records to {filepath}")
    else:
        logger.warning("No data fetched. Configure a data source.")

if __name__ == "__main__":
    main()
